educational_platform-main/requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2023.10
dj-database-url==2.3.0
Django==4.2.20
django-cleanup==8.0.0
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-widget-tweaks==1.5.0
gunicorn==21.2.0
packaging==25.0
pillow==10.2.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.0.0
sqlparse==0.5.3
typing_extensions==4.13.2
whitenoise==6.6.0
